Refactor index.js and displayItems, handleSearchInput functions
